Adds feature: "candlestick trend" plots
Adds a new boolean plot parameter called "trend" that can be used in conjunction with style='candle' in cerebro.plot. If invoked, the candlestick bars are colored and filled according to the "candlestick trend" definition (i.e. the color of the bar is selected based on the price action between the previous bar and the current bar, and the fill of the bar is selected by the price action of the current bar).

Default behavior: If "trend" parameter is not given, then 'style=candle' will default to the existing backtrader behavior for candlestick plots (i.e. trend=False)